Add files via upload
This is a Crime Reporting System that makes it easy for the complainer to register the crime and a central database that has hierarchial access based in the ranking of the officer. Such a central database will help the central government and its forces tackle situations and remove bureaucratic hurdles with the state police departments in a large country like India. Salient features include high scalability, quick response time and measures,  user friendly UI for the end user. Such quick response systems are not available in India, certainly not a MAKE IN INDIA Crime Reporting System.